+++ packages/stream_chat_v1/lib/channel_list_page.dart
@@ -1,195 +1,277 @@
 import 'dart:async';
 
 import 'package:example/routes/routes.dart';
-import 'package:example/search_text_field.dart';
+import 'package:example/user_mentions_page.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_app_badger/flutter_app_badger.dart';
+import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
+import 'package:streaming_shared_preferences/streaming_shared_preferences.dart';
 
-import 'channel_page.dart';
-import 'chat_info_screen.dart';
-import 'group_info_screen.dart';
+import 'channel_list.dart';
 
 class ChannelListPage extends StatefulWidget {
+  const ChannelListPage({
+    Key? key,
+  }) : super(key: key);
+
   @override
   _ChannelListPageState createState() => _ChannelListPageState();
 }
 
 class _ChannelListPageState extends State<ChannelListPage> {
-  TextEditingController? _controller;
+  int _currentIndex = 0;
 
-  String _channelQuery = '';
+  bool _isSelected(int index) => _currentIndex == index;
 
-  bool _isSearchActive = false;
-
-  Timer? _debounce;
-
-  void _channelQueryListener() {
-    if (_debounce?.isActive ?? false) _debounce!.cancel();
-    _debounce = Timer(const Duration(milliseconds: 350), () {
-      if (mounted) {
-        setState(() {
-          _channelQuery = _controller!.text;
-          _isSearchActive = _channelQuery.isNotEmpty;
-        });
-      }
-    });
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = TextEditingController()..addListener(_channelQueryListener);
-  }
-
-  @override
-  void dispose() {
-    _controller?.removeListener(_channelQueryListener);
-    _controller?.dispose();
-    super.dispose();
+  List<BottomNavigationBarItem> get _navBarItems {
+    return <BottomNavigationBarItem>[
+      BottomNavigationBarItem(
+        icon: Stack(
+          clipBehavior: Clip.none,
+          children: [
+            StreamSvgIcon.message(
+              color: _isSelected(0)
+                  ? StreamChatTheme.of(context).colorTheme.black
+                  : Colors.grey,
+            ),
+            Positioned(
+              top: -3,
+              right: -16,
+              child: UnreadIndicator(),
+            ),
+          ],
+        ),
+        label: 'Chats',
+      ),
+      BottomNavigationBarItem(
+        icon: Stack(
+          clipBehavior: Clip.none,
+          children: [
+            StreamSvgIcon.mentions(
+              color: _isSelected(1)
+                  ? StreamChatTheme.of(context).colorTheme.black
+                  : Colors.grey,
+            ),
+          ],
+        ),
+        label: 'Mentions',
+      ),
+    ];
   }
 
   @override
   Widget build(BuildContext context) {
     final user = StreamChat.of(context).user;
-    return WillPopScope(
-      onWillPop: () async {
-        if (_isSearchActive) {
-          _controller!.clear();
-          setState(() => _isSearchActive = false);
-          return false;
+    if (user == null) {
+      return Offstage();
+    }
+    return Scaffold(
+      backgroundColor: StreamChatTheme.of(context).colorTheme.whiteSnow,
+      appBar: ChannelListHeader(
+        onNewChatButtonTap: () {
+          Navigator.pushNamed(context, Routes.NEW_CHAT);
+        },
+        preNavigationCallback: () {
+          FocusScope.of(context).requestFocus(FocusNode());
+        },
+      ),
+      drawer: LeftDrawer(
+        user: user,
+      ),
+      drawerEdgeDragWidth: 50,
+      bottomNavigationBar: BottomNavigationBar(
+        backgroundColor: StreamChatTheme.of(context).colorTheme.white,
+        currentIndex: _currentIndex,
+        items: _navBarItems,
+        selectedLabelStyle: StreamChatTheme.of(context).textTheme.footnoteBold,
+        unselectedLabelStyle:
+            StreamChatTheme.of(context).textTheme.footnoteBold,
+        type: BottomNavigationBarType.fixed,
+        selectedItemColor: StreamChatTheme.of(context).colorTheme.black,
+        unselectedItemColor: Colors.grey,
+        onTap: (index) {
+          setState(() => _currentIndex = index);
+        },
+      ),
+      body: IndexedStack(
+        index: _currentIndex,
+        children: [
+          ChannelList(),
+          UserMentionsPage(),
+        ],
+      ),
+    );
+  }
+
+  StreamSubscription<int>? badgeListener;
+
+  @override
+  void initState() {
+    if (!kIsWeb) {
+      badgeListener = StreamChat.of(context)
+          .client
+          .state
+          .totalUnreadCountStream
+          .listen((count) {
+        if (count > 0) {
+          FlutterAppBadger.updateBadgeCount(count);
+        } else {
+          FlutterAppBadger.removeBadge();
         }
-        return true;
-      },
-      child: ChannelsBloc(
-        child: MessageSearchBloc(
-          child: NestedScrollView(
-            floatHeaderSlivers: true,
-            headerSliverBuilder: (_, __) => [
-              SliverToBoxAdapter(
-                child: SearchTextField(
-                  controller: _controller,
-                  showCloseButton: _isSearchActive,
-                ),
-              ),
-            ],
-            body: AnimatedSwitcher(
-              duration: const Duration(milliseconds: 350),
-              child: GestureDetector(
-                behavior: HitTestBehavior.opaque,
-                onPanDown: (_) => FocusScope.of(context).unfocus(),
-                child: _isSearchActive
-                    ? MessageSearchListView(
-                        showErrorTile: true,
-                        messageQuery: _channelQuery,
-                        filters: Filter.in_('members', [user!.id]),
-                        sortOptions: [
-                          SortOption(
-                            'created_at',
-                            direction: SortOption.ASC,
+      });
+    }
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    badgeListener?.cancel();
+    super.dispose();
+  }
+}
+
+class LeftDrawer extends StatelessWidget {
+  const LeftDrawer({
+    Key? key,
+    required this.user,
+  }) : super(key: key);
+
+  final User user;
+
+  @override
+  Widget build(BuildContext context) {
+    return Drawer(
+      child: Container(
+        color: StreamChatTheme.of(context).colorTheme.white,
+        child: SafeArea(
+          child: Padding(
+            padding: EdgeInsets.only(
+              top: MediaQuery.of(context).viewPadding.top + 8,
+            ),
+            child: Column(
+              children: [
+                Padding(
+                  padding: const EdgeInsets.only(
+                    bottom: 20.0,
+                    left: 8,
+                  ),
+                  child: Row(
+                    children: [
+                      UserAvatar(
+                        user: user,
+                        showOnlineStatus: false,
+                        constraints: BoxConstraints.tight(Size.fromRadius(20)),
+                      ),
+                      Padding(
+                        padding: const EdgeInsets.only(left: 16.0),
+                        child: Text(
+                          user.name,
+                          style: TextStyle(
+                            fontSize: 16,
+                            fontWeight: FontWeight.bold,
                           ),
-                        ],
-                        pullToRefresh: false,
-                        paginationParams: PaginationParams(limit: 20),
-                        emptyBuilder: (_) {
-                          return LayoutBuilder(
-                            builder: (context, viewportConstraints) {
-                              return SingleChildScrollView(
-                                physics: AlwaysScrollableScrollPhysics(),
-                                child: ConstrainedBox(
-                                  constraints: BoxConstraints(
-                                    minHeight: viewportConstraints.maxHeight,
-                                  ),
-                                  child: Center(
-                                    child: Column(
-                                      children: [
-                                        Padding(
-                                          padding: const EdgeInsets.all(24),
-                                          child: StreamSvgIcon.search(
-                                            size: 96,
-                                            color: Colors.grey,
-                                          ),
-                                        ),
-                                        Text(
-                                          'No results...',
-                                        ),
-                                      ],
-                                    ),
-                                  ),
-                                ),
-                              );
-                            },
-                          );
-                        },
-                        onItemTap: (messageResponse) async {
-                          FocusScope.of(context).requestFocus(FocusNode());
-                          final client = StreamChat.of(context).client;
-                          final message = messageResponse.message;
-                          final channel = client.channel(
-                            messageResponse.channel!.type,
-                            id: messageResponse.channel!.id,
-                          );
-                          if (channel.state == null) {
-                            await channel.watch();
-                          }
-                          Navigator.pushNamed(
-                            context,
-                            Routes.CHANNEL_PAGE,
-                            arguments: ChannelPageArgs(
-                              channel: channel,
-                              initialMessage: message,
-                            ),
-                          );
-                        },
-                      )
-                    : ChannelListView(
-                        onStartChatPressed: () {
-                          Navigator.pushNamed(context, Routes.NEW_CHAT);
-                        },
-                        swipeToAction: true,
-                        filter: Filter.in_('members', [user!.id]),
-                        presence: true,
-                        pagination: PaginationParams(
-                          limit: 20,
                         ),
-                        channelWidget: ChannelPage(),
-                        onViewInfoTap: (channel) {
-                          Navigator.pop(context);
-                          if (channel.memberCount == 2 && channel.isDistinct) {
-                            Navigator.push(
-                              context,
-                              MaterialPageRoute(
-                                builder: (context) => StreamChannel(
-                                  channel: channel,
-                                  child: ChatInfoScreen(
-                                    messageTheme: StreamChatTheme.of(context)
-                                        .ownMessageTheme,
-                                    user: channel.state!.members
-                                        .where((m) =>
-                                            m.userId !=
-                                            channel.client.state.user!.id)
-                                        .first
-                                        .user,
-                                  ),
-                                ),
-                              ),
-                            );
-                          } else {
-                            Navigator.push(
-                              context,
-                              MaterialPageRoute(
-                                builder: (context) => StreamChannel(
-                                  channel: channel,
-                                  child: GroupInfoScreen(
-                                    messageTheme: StreamChatTheme.of(context)
-                                        .ownMessageTheme,
-                                  ),
-                                ),
-                              ),
-                            );
-                          }
+                      ),
+                    ],
+                  ),
+                ),
+                ListTile(
+                  leading: StreamSvgIcon.penWrite(
+                    color: StreamChatTheme.of(context)
+                        .colorTheme
+                        .black
+                        .withOpacity(.5),
+                  ),
+                  onTap: () {
+                    Navigator.popAndPushNamed(
+                      context,
+                      Routes.NEW_CHAT,
+                    );
+                  },
+                  title: Text(
+                    'New direct message',
+                    style: TextStyle(
+                      fontSize: 14.5,
+                    ),
+                  ),
+                ),
+                ListTile(
+                  leading: StreamSvgIcon.contacts(
+                    color: StreamChatTheme.of(context)
+                        .colorTheme
+                        .black
+                        .withOpacity(.5),
+                  ),
+                  onTap: () {
+                    Navigator.popAndPushNamed(
+                      context,
+                      Routes.NEW_GROUP_CHAT,
+                    );
+                  },
+                  title: Text(
+                    'New group',
+                    style: TextStyle(
+                      fontSize: 14.5,
+                    ),
+                  ),
+                ),
+                Expanded(
+                  child: Container(
+                    alignment: Alignment.bottomCenter,
+                    child: ListTile(
+                      onTap: () async {
+                        Navigator.pop(context);
+
+                        if (!kIsWeb) {
+                          final secureStorage = FlutterSecureStorage();
+                          await secureStorage.deleteAll();
+                        }
+
+                        final client = StreamChat.of(context).client;
+                        client.disconnectUser();
+                        await client.dispose();
+
+                        await Navigator.of(
+                          context,
+                          rootNavigator: true,
+                        ).pushNamedAndRemoveUntil(
+                          Routes.CHOOSE_USER,
+                          ModalRoute.withName(Routes.CHOOSE_USER),
+                        );
+                      },
+                      leading: StreamSvgIcon.user(
+                        color: StreamChatTheme.of(context)
+                            .colorTheme
+                            .black
+                            .withOpacity(.5),
+                      ),
+                      title: Text(
+                        'Sign out',
+                        style: TextStyle(
+                          fontSize: 14.5,
+                        ),
+                      ),
+                      trailing: IconButton(
+                        icon: StreamSvgIcon.iconMoon(
+                          size: 24,
+                        ),
+                        color: StreamChatTheme.of(context).colorTheme.grey,
+                        onPressed: () async {
+                          final sp = await StreamingSharedPreferences.instance;
+                          sp.setInt(
+                            'theme',
+                            Theme.of(context).brightness == Brightness.dark
+                                ? 1
+                                : -1,
+                          );
                         },
                       ),
-              ),
+                    ),
+                  ),
+                ),
+              ],
             ),
           ),
         ),
