+++ packages/stream_chat_v1/lib/choose_user_page.dart
@@ -1,4 +1,5 @@
 import 'package:example/app_config.dart';
+import 'package:example/home_page.dart';
 import 'package:example/stream_version.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -6,6 +7,7 @@ import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
+import 'main.dart';
 import 'routes/routes.dart';
 
 const kStreamApiKey = 'STREAM_API_KEY';
@@ -89,8 +91,11 @@ class ChooseUserPage extends StatelessWidget {
                               ),
                             );
 
-                            final client = StreamChat.of(context).client;
-                            // client.apiKey = kDefaultStreamApiKey;
+                            final client = StreamChatClient(
+                              kDefaultStreamApiKey,
+                              logLevel: Level.INFO,
+                            )..chatPersistenceClient = chatPersistentClient;
+
                             await client.connectUser(
                               user,
                               token,
@@ -115,6 +120,7 @@ class ChooseUserPage extends StatelessWidget {
                               context,
                               Routes.HOME,
                               ModalRoute.withName(Routes.HOME),
+                              arguments: HomePageArgs(client),
                             );
                           },
                           leading: UserAvatar(
